Limit the fields returned to reduce API message size
Should make JSON.parse a lot faster to cut the response size
in half.  We should look at pagination next maybe.

Also did some url parsing action on the url so we can make it easier
to genericize in the future, and so we don't break the reducer if we
add more query params to the query.